Adapt code for DX 24.1.x
