+++ DXUploadXAFBlazor.Blazor.Server/DXUploadXAFBlazor.Blazor.Server.csproj
@@ -25,9 +25,7 @@
     <PackageReference Include="DevExpress.ExpressApp.Xpo" Version="24.1.5" />
     <PackageReference Include="DevExpress.Persistent.Base" Version="24.1.5" />
     <PackageReference Include="DevExpress.Persistent.BaseImpl.Xpo" Version="24.1.5" />
-    <PackageReference Include="DevExpress.ExpressApp" Version="24.1.5" />
     <PackageReference Include="DevExpress.ExpressApp.WebApi" Version="24.1.5" />
-    <PackageReference Include="DevExpress.ExpressApp.Xpo" Version="24.1.5" />
     <PackageReference Include="DevExpress.Persistent.Base" Version="24.1.5" />
     <PackageReference Include="DevExpress.Persistent.BaseImpl.Xpo" Version="24.1.5" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
