+++ DXUploadXAFBlazor.Module.Blazor/Editors/UploadFilePropertyEditor.cs
@@ -3,16 +3,13 @@ using DevExpress.ExpressApp.Blazor.Editors;
 using DevExpress.ExpressApp.Blazor.Editors.Adapters;
 using DevExpress.ExpressApp.Editors;
 using DevExpress.ExpressApp.Model;
-using DevExpress.Persistent.Base;
 using DevExpress.Persistent.BaseImpl;
+using DXUploadXAFBlazor.Module.Blazor.Middlewares;
 using DXUploadXAFBlazor.Module.BusinessObjects;
-using Microsoft.AspNetCore.Http;
 using System;
 using System.IO;
 using System.Linq;
 using System.Reactive.Linq;
-using System.Reactive.Subjects;
-using System.Threading.Tasks;
 
 namespace DXUploadXAFBlazor.Module.Blazor.Editors
 {
@@ -38,7 +35,7 @@ namespace DXUploadXAFBlazor.Module.Blazor.Editors
                 {
                     var fileData = (FileData)objectSpace.CreateObject(typeof(FileData));
                     fileData.LoadFromStream(formFile.name, new MemoryStream(formFile.bytes));
-                    
+
                     ((DomainObject1)CurrentObject).Data = fileData;
                     ((DomainObject1)CurrentObject).FileData = fileData;
                     objectSpace.CommitChanges();
@@ -47,37 +44,4 @@ namespace DXUploadXAFBlazor.Module.Blazor.Editors
                 .Subscribe();
         }
     }
-
-    public class UploadFileMiddleware
-    {
-        static readonly ISubject<(string name, byte[] bytes, string editor)> FormFileSubject = Subject.Synchronize(new Subject<(string name, byte[] bytes, string editor)>());
-
-        public static IObservable<(string name, byte[] bytes, string editor)> FormFile => FormFileSubject.AsObservable();
-
-        private readonly RequestDelegate _next;
-        public UploadFileMiddleware(RequestDelegate next)
-        {
-            this._next = next;
-        }
-        public static byte[] streamToByteArray(Stream input)
-        {
-            MemoryStream ms = new MemoryStream();
-            input.CopyTo(ms);
-            return ms.ToArray();
-        }
-        public async System.Threading.Tasks.Task Invoke(HttpContext context)
-        {
-            string requestPath = context.Request.Path.Value.TrimStart('/');
-            if (requestPath.StartsWith("api/Upload/UploadFile"))
-            {
-                var formFile = context.Request.Form.Files.First();
-                FormFileSubject.OnNext((formFile.FileName, streamToByteArray(formFile.OpenReadStream()), context.Request.Query["Editor"]));
-            }
-            else
-            {
-                await _next(context);
-            }
-        }
-    }
 }
-
